refactor: detail 이름 복구
