+++ src/components/detail/MenuDetail.tsx
